Move Swagger generation inside a special type of test
- Add a new configuration for failsafe plugin to hanle *ITSupport.java which will generate the swagger file
- Create a profile generate-swagger enabled by default (can be skipped with `-DskipSwaggerGen` option)
